feat: moved structs to specific commands
