+++ internal/pkg/service/cli/cmd/dbt/init.go
@@ -9,6 +9,12 @@ import (
 	initOp "github.com/keboola/keboola-as-code/pkg/lib/operation/dbt/init"
 )
 
+type DbtFlags struct {
+	StorageAPIHost string `mapstructure:"storage-api-host" shorthand:"H" usage:"storage API host, eg. \"connection.keboola.com\""`
+	TargetName     string `mapstructure:"target-name" shorthand:"T" usage:"target name of the profile"`
+	WorkspaceName  string `mapstructure:"workspace-name" shorthand:"W" usage:"name of workspace to create"`
+}
+
 func InitCommand(p dependencies.Provider) *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   `init`,
@@ -36,8 +42,7 @@ func InitCommand(p dependencies.Provider) *cobra.Command {
 		},
 	}
 
-	initFlags := Flags{}
-	_ = cliconfig.GenerateFlags(initFlags, cmd.Flags())
+	cliconfig.MustGenerateFlags(DbtFlags{}, cmd.Flags())
 
 	return cmd
 }
