+++ internal/pkg/service/cli/cmd/remote/table/unload.go
@@ -15,6 +15,28 @@ import (
 	"github.com/keboola/keboola-as-code/pkg/lib/operation/project/remote/table/unload"
 )
 
+type UnloadFlags struct {
+	StorageAPIHost string   `mapstructure:"storage-api-host" shorthand:"H" usage:"storage API host, eg. \"connection.keboola.com\""`
+	ChangedSince   string   `mapstructure:"changed-since" usage:"only export rows imported after this date"`
+	ChangedUntil   string   `mapstructure:"changed-until" usage:"only export rows imported before this date"`
+	Columns        []string `mapstructure:"columns" usage:"comma-separated list of columns to export"`
+	Limit          uint     `mapstructure:"limit" usage:"limit the number of exported rows"`
+	Where          string   `mapstructure:"where" usage:"filter columns by value"`
+	Order          string   `mapstructure:"order" usage:"order by one or more columns"`
+	Format         string   `mapstructure:"format" usage:"output format (json/csv)"`
+	Async          bool     `mapstructure:"async" usage:"do not wait for unload to finish"`
+	Timeout        string   `mapstructure:"timeout" usage:"how long to wait for job to finish"`
+}
+
+func DefaultUnloadFlags() *UnloadFlags {
+	return &UnloadFlags{
+		Limit:   0,
+		Columns: []string{},
+		Format:  "csv",
+		Timeout: "5m",
+	}
+}
+
 func UnloadCommand(p dependencies.Provider) *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   `unload [table]`,
@@ -56,13 +78,7 @@ func UnloadCommand(p dependencies.Provider) *cobra.Command {
 		},
 	}
 
-	unloadFlags := UnloadFlags{
-		Limit:   0,
-		Columns: []string{},
-		Format:  "csv",
-		Timeout: "5m",
-	}
-	_ = cliconfig.GenerateFlags(unloadFlags, cmd.Flags())
+	cliconfig.MustGenerateFlags(DefaultUnloadFlags(), cmd.Flags())
 
 	return cmd
 }
