+++ internal/pkg/service/cli/cmd/sync/init.go
@@ -11,6 +11,27 @@ import (
 	initOp "github.com/keboola/keboola-as-code/pkg/lib/operation/project/sync/init"
 )
 
+type InitFlags struct {
+	StorageAPIHost string `mapstructure:"storage-api-host" shorthand:"H" usage:"storage API host, eg. \"connection.keboola.com\""`
+	Branches       string `mapstructure:"branches" shorthand:"b" usage:"comma separated IDs or name globs, use \"*\" for all"`
+	CI             bool   `mapstructure:"ci" usage:"generate workflows"`
+	CIValidate     bool   `mapstructure:"ci-validate" usage:"create workflow to validate all branches on change"`
+	CIPush         bool   `mapstructure:"ci-push" usage:"create workflow to push change in main branch to the project"`
+	CIPull         bool   `mapstructure:"ci-pull" usage:"create workflow to sync main branch each hour"`
+	CIMainBranch   string `mapstructure:"ci-main-branch" usage:"name of the main branch for push/pull workflows"`
+}
+
+func DefaultInitFlags() *InitFlags {
+	return &InitFlags{
+		CI:           true,
+		CIValidate:   true,
+		CIPull:       true,
+		CIPush:       true,
+		CIMainBranch: "main",
+		Branches:     "main",
+	}
+}
+
 func InitCommand(p dependencies.Provider) *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   "init",
@@ -42,15 +63,7 @@ func InitCommand(p dependencies.Provider) *cobra.Command {
 		},
 	}
 
-	initFlags := InitFlags{
-		CI:           true,
-		CIValidate:   true,
-		CIPull:       true,
-		CIPush:       true,
-		CIMainBranch: "main",
-		Branches:     "main",
-	}
-	_ = cliconfig.GenerateFlags(initFlags, cmd.Flags())
+	cliconfig.MustGenerateFlags(DefaultInitFlags(), cmd.Flags())
 
 	return cmd
 }
