+++ internal/pkg/service/cli/cmd/template/create.go
@@ -9,6 +9,18 @@ import (
 	createOp "github.com/keboola/keboola-as-code/pkg/lib/operation/template/local/create"
 )
 
+type CreateFlags struct {
+	ID             string `mapstructure:"id" usage:"template ID"`
+	Name           string `mapstructure:"name" usage:"template name"`
+	Description    string `mapstructure:"description" usage:"template description"`
+	StorageAPIHost string `mapstructure:"storage-api-host" shorthand:"H" usage:"storage API host, eg. \"connection.keboola.com\""`
+	Branch         string `mapstructure:"branch" shorthand:"b" usage:"branch ID or name"`
+	Configs        string `mapstructure:"configs" shorthand:"c" usage:"comma separated list of {componentId}:{configId}"`
+	UsedComponents string `mapstructure:"used-components" shorthand:"u" usage:"comma separated list of component ids"`
+	AllConfigs     bool   `mapstructure:"all-configs" shorthand:"a" usage:"use all configs from the branch"`
+	AllInputs      bool   `mapstructure:"all-inputs" usage:"use all found config/row fields as user inputs"`
+}
+
 func CreateCommand(p dependencies.Provider) *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   "create",
@@ -32,8 +44,7 @@ func CreateCommand(p dependencies.Provider) *cobra.Command {
 		},
 	}
 
-	createFlags := CreateFlags{}
-	_ = cliconfig.GenerateFlags(createFlags, cmd.Flags())
+	cliconfig.MustGenerateFlags(CreateFlags{}, cmd.Flags())
 
 	return cmd
 }
